SYNCT-211: Create ExtensionsUtil
